✅ Move all passed test to passed_test.go
