+++ tests/conditional_test/passed_test.go
@@ -152,3 +152,187 @@ func TestCWLWhen_RunIdx19(t *testing.T) {
 	t.Log("Doc: ", test.Doc)
 	doTest(t, test)
 }
+
+func TestCWLWhen_RunIdx20(t *testing.T) {
+	test := tests[20]
+	t.Log("Test 20: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx21(t *testing.T) {
+	test := tests[21]
+	t.Log("Test 21: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx22(t *testing.T) {
+	test := tests[22]
+	t.Log("Test 22: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx23(t *testing.T) {
+	test := tests[23]
+	t.Log("Test 23: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx24(t *testing.T) {
+	test := tests[24]
+	t.Log("Test 24: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx25(t *testing.T) {
+	test := tests[25]
+	t.Log("Test 25: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx26(t *testing.T) {
+	test := tests[26]
+	t.Log("Test 26: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx27(t *testing.T) {
+	test := tests[27]
+	t.Log("Test 27: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx28(t *testing.T) {
+	test := tests[28]
+	t.Log("Test 28: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx29(t *testing.T) {
+	test := tests[29]
+	t.Log("Test 29: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+
+func TestCWLWhen_RunIdx30(t *testing.T) {
+	test := tests[30]
+	t.Log("Test 30: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx31(t *testing.T) {
+	test := tests[31]
+	t.Log("Test 31: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx32(t *testing.T) {
+	test := tests[32]
+	t.Log("Test 32: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx33(t *testing.T) {
+	test := tests[33]
+	t.Log("Test 33: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx34(t *testing.T) {
+	test := tests[34]
+	t.Log("Test 34: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx35(t *testing.T) {
+	test := tests[35]
+	t.Log("Test 35: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx36(t *testing.T) {
+	test := tests[36]
+	t.Log("Test 36: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx37(t *testing.T) {
+	test := tests[37]
+	t.Log("Test 37: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx38(t *testing.T) {
+	test := tests[38]
+	t.Log("Test 38: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx39(t *testing.T) {
+	test := tests[39]
+	t.Log("Test 39: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx40(t *testing.T) {
+	test := tests[40]
+	t.Log("Test 40: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx41(t *testing.T) {
+	test := tests[41]
+	t.Log("Test 41: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx42(t *testing.T) {
+	test := tests[42]
+	t.Log("Test 42: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx43(t *testing.T) {
+	test := tests[43]
+	t.Log("Test 43: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx44(t *testing.T) {
+	test := tests[44]
+	t.Log("Test 44: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
+func TestCWLWhen_RunIdx45(t *testing.T) {
+	test := tests[45]
+	t.Log("Test 45: ")
+	t.Log("ID : ", test.ID)
+	t.Log("Doc: ", test.Doc)
+	doTest(t, test)
+}
